requirements.txt
python-telegram-bot==13.15
transformers==4.39.3
torch
